+++ src/desktop/Designer/pages/DesignerPage.tsx
@@ -1,8 +1,9 @@
 import { css } from '@emotion/react';
-import { Link } from 'react-router-dom';
 import { colors, fonts } from '../../../styles/theme';
 import { ImgBg3Web } from '../../assets/image';
 import PageLayout from '../../Common/PageLayout';
+import DesignerContact from '../components/DesignerContact';
+import DesignerWorks from '../components/DesignerWorks';
 
 const DUMMY = {
   name: '양종욱',
@@ -46,34 +47,6 @@ const DUMMY = {
 const DesignerPage = () => {
   const { name, engName, major, email, instagram, behance, works } = DUMMY;
 
-  const updateStudioUrl = (studioNm: string) => {
-    switch (studioNm) {
-      case '시각디자인스튜디오 A':
-        return '/studio-a';
-
-      case '시각디자인스튜디오 B':
-        return '/studio-b';
-
-      case '디자인씽킹스튜디오':
-        return '/design-thinking';
-
-      case '정보경험디자인스튜디오':
-        return '/user-experience';
-
-      case '모션그래픽스스튜디오':
-        return '/motion-graphics';
-
-      case '공간디자인스튜디오':
-        return '/space';
-
-      case '공간연출디자인스튜디오':
-        return '/space-direction';
-
-      default:
-        return '/';
-    }
-  };
-
   return (
     <PageLayout>
       <img src={ImgBg3Web} css={bg} />
@@ -86,42 +59,12 @@ const DesignerPage = () => {
         <span css={mainMajor}>{major}</span>
 
         <article css={additionalInfo}>
-          <div css={contactContainer}>
-            <div css={contact}>
-              <p css={contactCategory}>E-MAIL</p>
-              <p css={contactContents}>{email}</p>
-            </div>
-            {instagram && (
-              <div css={contact}>
-                <p css={contactCategory}>INSTAGRAM</p>
-                <p css={contactContents}>{instagram}</p>
-              </div>
-            )}
-            {behance && (
-              <div css={contact}>
-                <p css={contactCategory}>BEHANCE</p>
-                <p css={contactContents}>{behance}</p>
-              </div>
-            )}
-          </div>
-
-          <div css={worksContainer}>
-            {works.map((work) => {
-              const { url, workTitle, studioNm, images } = work;
-              const { imgPath } = images.length > 1 ? images[1] : images[0];
-              const studioUrl = updateStudioUrl(studioNm);
-
-              return (
-                <Link to={`${studioUrl}/${url}`}>
-                  <img src={imgPath} css={workImg} />
-                  <div css={workInfoContainer}>
-                    <p css={title}>{workTitle}</p>
-                    <p css={studioName}>{studioNm}</p>
-                  </div>
-                </Link>
-              );
-            })}
-          </div>
+          <DesignerContact
+            email={email}
+            instagram={instagram}
+            behance={behance}
+          />
+          <DesignerWorks works={works} />
         </article>
       </section>
     </PageLayout>
@@ -175,57 +118,4 @@ const additionalInfo = css`
   align-items: flex-start;
 
   margin-top: 11.8rem;
-`;
-
-const contactContainer = css`
-  display: flex;
-  gap: 2rem;
-  justify-content: center;
-  flex-direction: column;
-`;
-
-const contact = css`
-  display: flex;
-  gap: 0.4rem;
-  justify-content: center;
-  flex-direction: column;
-`;
-
-const contactCategory = css`
-  color: ${colors.gray300};
-  ${fonts.desktop_body_semi_20};
-`;
-
-const contactContents = css`
-  color: ${colors.gray900};
-  ${fonts.desktop_body_reg_20};
-`;
-
-const worksContainer = css`
-  display: flex;
-  gap: 3.2rem;
-  align-items: center;
-`;
-
-const workImg = css`
-  width: 30.6rem;
-  height: 30.6rem;
-`;
-
-const workInfoContainer = css`
-  display: flex;
-  gap: 0.4rem;
-  flex-direction: column;
-
-  margin-top: 1.2rem;
-`;
-
-const title = css`
-  color: ${colors.gray900};
-  ${fonts.desktop_body_semi_20};
-`;
-
-const studioName = css`
-  color: ${colors.gray900};
-  ${fonts.desktop_body_reg_18};
 `;
